fix: 중복 경로 파일 통합 #79
