+++ src/app/(root)/(mainpage)/globals.css
