+++ src/app/(root)/(mainpage)/page.tsx
@@ -5,7 +5,6 @@ import AllPage from "./all/page";
 const MainPage = async () => {
   return (
     <div className="w-full mx-auto max-w-container-l m:max-w-container-m s:max-w-container-s">
-      <NavTabs />
       <AllPage />
     </div>
   );
