--- src/app/(root)/page.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import MainPage from "./(mainpage)/page";
-
-export default function Home() {
-  return <MainPage />;
-}
